+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ### 1.8.0
  - Added support for for loops variable completion.
  - Added support for enums parsing.
- - Added a command for tasks.json generation from template.
- - Added a command for .sp file generation from template.
+ - Added a command for tasks.json generation from a template.
+ - Added a command for .sp file generation from a template.
+ - Added a command for README.md file generation from a template.
  - Added more settings to support the above features.
 
 ### 1.7.1